move machines to its own module
