+++ package.yaml
@@ -61,6 +61,8 @@ library:
       - Upcast.IO
       - Upcast.Infra
       - Upcast.Infra.Amazonka
+      - Upcast.Infra.AmazonkaTypes
+      - Upcast.Infra.Machine
       - Upcast.Infra.Match
       - Upcast.Infra.NixTypes
       - Upcast.Infra.Types
